# db-files/fast-api-generate/app/models.py
#
# Copyright (c) 2023 by Delphix. All rights reserved.
#

# generated by fastapi-codegen:
#   filename:  api.yaml
#   timestamp: 2023-06-01T05:37:27+00:00

from __future__ import annotations

from enum import Enum
from typing import Dict, List, Optional

from pydantic import BaseModel, Field, constr


class ConnectorResponse(BaseModel):
    id: Optional[int] = Field(
        None, description="The Connector object entity ID."
    )  # noqa
    user: constr(min_length=1, max_length=256) = Field(
        ..., description="The username of this connector."
    )
    jdbc_url: constr(min_length=1, max_length=256) = Field(
        ..., description="The jdbc url of this connector."
    )
    connection_properties: Optional[Dict[str, str]] = Field(
        None, description="The properties of this connector."
    )


class Connector(ConnectorResponse):
    password: constr(min_length=1, max_length=256) = Field(
        ..., description="The password of this connector."
    )
    restoreSensitiveFields: Optional[bool] = Field(
        False,
        description="Passed to indicate that user want to restore the sensitive data.",  # noqa
    )
    jdbc_url: constr(min_length=1, max_length=256) = Field(
        ..., description="The jdbc url of this connector."
    )
    user: constr(min_length=1, max_length=256) = Field(
        ..., description="The username of this connector."
    )


class DataInfoItem(BaseModel):
    schema_name: constr(min_length=1) = Field(
        ..., description="Database Schema name of source data."
    )
    table_name: constr(min_length=1) = Field(
        ..., description="Database Table name of source data."
    )
    filter_key: Optional[constr(min_length=1)] = Field(
        None,
        description="The unique database column field to filter the source data.",  # noqa
    )
    unload_split: Optional[int] = Field(
        1,
        description="The number of unloaded files to be generated from the source database.",  # noqa
    )


class DataSet(BaseModel):
    id: Optional[int] = None
    connector_id: int = Field(
        ...,
        description="ID of the Connector model to be used for this Data Set.",  # noqa
    )
    mount_filesystem_id: int = Field(
        ...,
        description="ID of the Mount Filesystem model to be used for this Data Set.",  # noqa
    )
    data_info: List[DataInfoItem] = Field(
        ...,
        description="Array of multiple objects, each providing the information of the table, schema etc.",  # noqa
        max_items=10000,
        min_items=1,
    )


class SourceConfigs(BaseModel):
    max_concurrent_source_connection: Optional[int] = Field(
        None,
        description="Maximum number of parallel connection that hyperscale can have with source datasource.",  # noqa
    )


class Unload(BaseModel):
    execution_id: int = Field(..., description="The Unload object entity ID.")
    job_id: int = Field(
        ..., description="ID of the Job model to be used for this Unload."
    )
    dataset_id: int = Field(
        ..., description="ID of the Data Set model to be used for this Unload."
    )
    source_configs: Optional[SourceConfigs] = Field(
        None, description="configuration properties for source datasource"
    )


class Status(Enum):
    CREATED = "CREATED"
    CANCELLED = "CANCELLED"
    FAILED = "FAILED"
    RUNNING = "RUNNING"
    SUCCEEDED = "SUCCEEDED"


class DataInfoItem1(BaseModel):
    source_key: Dict[str, str] = Field(
        ...,
        description="object providing the information like schema_name, table_name or file_name of the source datasource",  # noqa
    )
    rows_unloaded: int = Field(
        ...,
        description="Total number of rows unloaded from the database table.",  # noqa
    )
    unload_file_path: constr(min_length=1) = Field(
        ..., description="relative path of the unload source file."
    )
    status: Status = Field(..., description="The status of the unload process.")  # noqa
    error: Optional[str] = Field(None, description="Exception details.")


class Status1(Enum):
    CREATED = "CREATED"
    CANCELLED = "CANCELLED"
    FAILED = "FAILED"
    RUNNING = "RUNNING"
    SUCCEEDED = "SUCCEEDED"


class ExecutionStatus(BaseModel):
    status: Status1 = Field(
        ..., description="The status of the Unload process."
    )  # noqa
    error: Optional[str] = Field(None, description="Optional, error detail.")


class Status2(Enum):
    FAILED = "FAILED"
    RUNNING = "RUNNING"
    SUCCEEDED = "SUCCEEDED"
    CREATED = "CREATED"
    CANCELLED = "CANCELLED"


class SourceInfoItem(BaseModel):
    source_key: Dict[str, str] = Field(
        ...,
        description="Object providing the information like schema_name, table_name or file_name of the source datasource",  # noqa
    )
    unload_split_count: int = Field(
        ...,
        description="The number of unloaded files to be generated from the source database.",  # noqa
    )
    total_number_of_rows: int = Field(
        ...,
        description="Total number of rows unloaded from the database table.",  # noqa
    )
    masking_split_count: Optional[int] = Field(
        None, description="The number of masked splitted files to be generated."  # noqa
    )


class SourceData(BaseModel):
    status: Optional[Status2] = Field(
        None, description="The status of the Unload process."
    )
    source_info: Optional[List[SourceInfoItem]] = Field(
        None,
        description="Array of multiple objects, each object providing the information like sourceKey, unloadSplitCount etc.",  # noqa
        max_items=10000,
        min_items=1,
    )


class ApiVersion(BaseModel):
    versionId: constr(min_length=1, max_length=50) = Field(
        ...,
        description="The VersionId is used to return the current version of service.",  # noqa
    )


class UnloadResponse(ExecutionStatus):
    execution_id: int = Field(..., description="The Unload object entity ID.")
    job_id: int = Field(
        ..., description="ID of the Job model to be used for this Unload."
    )
    start_time: str = Field(
        ..., description="The time when the unload process is started."
    )
    end_time: Optional[str] = Field(
        None, description="The time when the unload process is completed."
    )
    data_info: List[DataInfoItem1] = Field(
        ...,
        description="Array of multiple objects, each object providing the information like sourceKey and file_info.",  # noqa
        max_items=10000,
        min_items=1,
    )